final changes of the day, more sutff done tomorrow (like crc re calculation/updation)

+++ src/chunk/chunk.cpp
@@ -48,7 +48,7 @@ string Chunk::GetType()
 	return _type;
 }
 
-int Chunk::GetCRC()
+unsigned long Chunk::GetCRC()
 {
 	return _cyclicRedundancyCheck;
 }
